discord-stat-bot-py/requirements.txt
discord.py==2.3.2
psutil